# vmmc backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/pre-employment/{employeeId}": {
      "get": {
        "operationId": "PreEmploymentController_get",
        "parameters": [
          {
            "name": "employeeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "pre-employment"
        ]
      },
      "put": {
        "operationId": "PreEmploymentController_upsert",
        "parameters": [
          {
            "name": "employeeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Client-generated key so a retried request never creates a duplicate record.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertPreEmploymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "pre-employment"
        ]
      }
    },
...
  },
  "info": {
    "title": "VMMC TB DOTS API",
    "description": "ITIL 4 Monitoring & Event Management backend for the VMMC TB DOTS health management system. Write endpoints accept an `Idempotency-Key` header so a client retry after a timeout never duplicates a record.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpsertPreEmploymentDto": {
        "type": "object",
        "properties": {
          "baselineXrayPath": {
            "type": "string",
            "description": "Storage reference for the baseline chest X-ray, if captured."
          },
          "screeningResult": {
            "type": "string",
            "description": "Free-text baseline screening result."
          },
          "fitForDuty": {
            "type": "boolean",
            "description": "Whether the employee has been cleared fit for duty."
          }
        }
      },
...
    }
  }
}